Move eligibility and service refinements code to component directory.
